third_party/WebKit/LayoutTests/web-animations-api/animation-id.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
    var animation = document.documentElement.animate([], { id: 'testId' });
    assert_equals(animation.id, 'testId');
}, 'Tests that animation id is set when specified from keyframe options');
</script>
